Crate icu_timezone_data

Data for the icu_timezone crate

This data was generated with CLDR version 45.0.0, ICU version icu4x/2024-05-16/75.x, and LSTM segmenter version v0.1.0.

Macros

impl_time_zone_bcp47_to_iana_v1

Implement DataProvider<Bcp47ToIanaMapV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_time_zone_iana_to_bcp47_v1

Implement DataProvider<IanaToBcp47MapV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_time_zone_iana_to_bcp47_v2

Implement DataProvider<IanaToBcp47MapV2Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_time_zone_metazone_period_v1

Implement DataProvider<MetazonePeriodV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impliterable_time_zone_bcp47_to_iana_v1

Implement IterableDataProvider<Bcp47ToIanaMapV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_time_zone_iana_to_bcp47_v1

Implement IterableDataProvider<IanaToBcp47MapV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_time_zone_iana_to_bcp47_v2

Implement IterableDataProvider<IanaToBcp47MapV2Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_time_zone_metazone_period_v1

Implement IterableDataProvider<MetazonePeriodV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

make_provider

Marks a type as a data provider. You can then use macros like impl_core_helloworld_v1 to add implementations.